+++ lib/src/network_image_badge_avatar.dart
@@ -3,8 +3,10 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 
-// [placeholder image] you just pass the asset name. For example : "assets/images/user_placeholder.png" or for string name use : Images.name. Else it will not show your image in placeholder.
-// By default we are showing one image from our root folder
+/// [placeholder image] you just pass the asset name. 
+/// For example : "assets/images/user_placeholder.png" or for string name use : Images.name. Else it will not show your image in placeholder.
+/// By default we are showing one image from our root folder
+/// 
 class NetworkImageBadgeAvatar extends StatelessWidget {
   final String? centeralText;
   final Color? bgColor;
@@ -48,7 +50,10 @@ class NetworkImageBadgeAvatar extends StatelessWidget {
             fit: StackFit.expand,
             alignment: Alignment.topRight,
             children: [
-              if (imagePath != null || imageString != null)             
+              if (imagePath != null 
+              || imageString != null
+              || centeralText != null
+              )             
                 Container(
                   decoration: BoxDecoration(
                     shape: BoxShape.circle,
@@ -58,46 +63,66 @@ class NetworkImageBadgeAvatar extends StatelessWidget {
                     ),
                   ),
                   child: Container(
-                    decoration: const BoxDecoration(
-                      color: Colors.transparent,
+                    decoration: BoxDecoration(
+                      color: bgColor ?? Colors.transparent,
                       shape: BoxShape.circle,
                     ),
                     clipBehavior: Clip.hardEdge,
                     alignment: Alignment.center,
                     child: SizedBox(
                       width: 180,
-                      child: Image.file(
-                        imagePath?.path != null
-                            ? File(imagePath!.path)
-                            : File(imageString!),
+                      child: imagePath?.path != null
+                      ? Image.file(                     
+                          File(imagePath!.path),                          
                         fit: BoxFit.cover,
+                      )
+                      : imageString != null            
+                      ? Image.file(                 
+                             File(imageString!),
+                        fit: BoxFit.cover,
+                      )
+                      : FittedBox(
+                        fit: BoxFit.scaleDown,
+                        child: Text(
+                          centeralText ?? "AI",
+                          style: TextStyle(
+                            fontSize: 24,
+                            color: textColor,
+                          ),
+                        ),
                       ),
                     ),
                   ),
-                )                           
+                )
               else
                 Container(
                   decoration: BoxDecoration(
-                    color: bgColor ?? const Color(0x00FFFFFF),
+                    /// [color] color: bgColor ?? const Color(0x00FFFFFF),
                     shape: BoxShape.circle,
                     border: Border.all(
                       color: borderColor ?? Colors.blueGrey,
                       width: 1.5,
                     ),
-                  ),
-                  padding: const EdgeInsets.all(8),
-                  clipBehavior: Clip.hardEdge,
-                  alignment: Alignment.center,
-                  child: FadeInImage.assetNetwork(
-                    placeholder: placeholderImage ??
-                        "assets/images/user_placeholder.png",
-                    image: networkImage ??
-                        'https://raw.githubusercontent.com/muktabd/public-images/main/user_placeholder.png',
-                    fit: BoxFit.cover,
+                  ),            
+                  child: Container(
+                    decoration: const BoxDecoration(
+                      color: Colors.transparent,
+                      shape: BoxShape.circle,
+                    ),
+                     clipBehavior: Clip.hardEdge,
+                    alignment: Alignment.center,
+                    child: FadeInImage.assetNetwork(
+                      placeholder: placeholderImage ??
+                          "assets/images/user_placeholder.png",
+                      image: networkImage ??
+                          'https://raw.githubusercontent.com/muktabd/public-images/main/user_placeholder.png',
+                      fit: BoxFit.cover,
+                    ),
                   ),
                 ),
                             
               /// adjustable iconbutton [image picker icon button]
+              /// by default top value set 85.0 based on SizedBox height 120
               Positioned(
                 top: iconPosition ?? 85.0,
                 right: 0,
@@ -122,6 +147,8 @@ class NetworkImageBadgeAvatar extends StatelessWidget {
                   ),
                 ),
               ),
+              ///
+              ///                 
             ],
           ),
         );