+++ CHANGELOG.md
@@ -1,6 +1,35 @@
-## 0.0.7
+## Curren version : 0.0.8
 
 * TODO: Describe initial release.
 There are three addintional widgets added to use.
+
 ## Main Changes in V0.0.7
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.
+
+## 0.0.6
+
+## Whats new in this version:
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.
+
+## 0.0.5
+
+## Whats new in this version:
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.
+
+## 0.0.4
+
+## Whats new in this version:
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.
+## 0.0.3
+
+## Whats new in this version:
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.
+## 0.0.2
+
+## Whats new in this version:
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.
+
+## 0.0.1
+
+## Whats new in this version:
 Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.